+++ app/modules/organization/profile/organization-profile.ctrl.js
@@ -1,40 +1,26 @@
 angular.module('organization')
-.controller('orgProfileCtrl', function(API,APIError,DataStorage,Loader,User,$scope,$state,$timeout) {
+.controller('orgProfileCtrl', function(DataStorage, Loader, Organization, User) {
 
     const orgProfile = this
-    const eventName = 'orgProfile.init'
+    const storedData = DataStorage.getType('orgProfile')
 
     orgProfile.organization = User.user.organization
-    orgProfile.organization ? APIError.offFor(eventName) : APIError.onFor(eventName)
 
     /* -------------------------------------------- ON LOAD START --------------------------------------------- */
 
-    const storedData = DataStorage.getType('orgProfile')
-
-    if (storedData) {
-        orgProfile.securityAdmins = storedData
+    if (storedData !== undefined) {
+        orgProfile.securityAdmins = storedData.admins
+        orgProfile.passwordPolicy = storedData.passwordPolicy
     }
+    else Loader.onFor('orgProfile.init')
 
-    if (storedData || orgProfile.organization) {
-        if (!storedData) Loader.onFor(eventName)
-
-        API.cui.getPersonsAdmins({qs: [['organization.id', orgProfile.organization.id], ['securityadmin', true]]})
-        .done(res => {
-            orgProfile.securityAdmins = res
-            DataStorage.setType('orgProfile', res)
-        })
-        .fail(err => {
-            console.error('Error getting organization admin information', err)
-            APIError.onFor(eventName)
-            $timeout(() => {
-                APIError.offFor(eventName)
-            }, 5000)
-        })
-        .always(() => {
-            Loader.offFor(eventName)
-            $scope.$digest()
-        })
-    }
+    Organization.initOrganizationProfile(orgProfile.organization.id, orgProfile.organization.passwordPolicy.id)
+    .then(res => {
+        orgProfile.securityAdmins = res.admins
+        orgProfile.passwordPolicy = res.passwordPolicy
+        DataStorage.setType('orgProfile', res)
+        Loader.offFor('orgProfile.init')
+    })
 
     /* --------------------------------------------- ON LOAD END ---------------------------------------------- */